Skip .appSourceCopPackages when cleanup apps

+++ PSScripts/NAV Extensions v2/MultiRootWorkspaces/Apps_App_Cleanup.ps1
@@ -2,8 +2,12 @@
 
 foreach ($Target in $Targets) {
     write-Host $Target
-    Get-ChildItem -Path $Target -Filter '*.app' -Recurse | % {
-        $_ | Remove-Item -Force -Verbose
+    Get-ChildItem -Path $Target -Filter '*.app' -Recurse -Exclude "*\.appSourceCopPackages\*" | % {
+        if ($_.FullName -like "*.appSourceCopPackages*"){
+            write-Host -ForegroundColor Yellow "Skipped $($_.FullName)"
+        } else {
+            $_ | Remove-Item -Force -Verbose
+        }
     }
 }
